% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flo_umap.R
\name{flo_umap}
\alias{flo_umap}
\title{This function makes pretty UMAP plots. It takes a dataframe as input and uses the UMAP1 and UMAP2 columns as X and Y coordinates and any
other named column as the coloring variable. It can facet by one variable.}
\usage{
flo_umap(df, color_by, facet_by = NULL, only_show = NULL)
}
\arguments{
\item{df;}{\code{data.frame} containing cells as rows and features as columns}

\item{color_by;}{character string; the coloring channel; has to be one of \code{colnames(df)}}

\item{facet_by;}{character string; the facetting variable; has to be NULL or one of \code{colnames(df)}}
}
\value{
a ggplot2 object
}
\description{
It relies on the ggplot2 package
}
\examples{
sce <- read_small("~/PhD/cytof/vac69a/reprocessed/reprocessed_relabeled_comped/T_cells_only/", 500)
df <- prep_sce_for_ggplot(sce)
cd25_plot <- flo_umap(df, "CD25", "timepoint")
}
