% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sce_for_ggplot.R
\name{prep_sce_for_ggplot}
\alias{prep_sce_for_ggplot}
\title{This function makes data.frames when passed SingleCellExperiments. This allows making ggplot objects from scratch using sce data.}
\usage{
prep_sce_for_ggplot(sce)
}
\arguments{
\item{sce;}{a SingleCellExperiment containing UMAP dimensions}
}
\value{
a data.frame containing events as rows and features as columns
}
\description{
It relies on the ggplot2 package
}
\examples{
big_table <- prep_sce_for_ggplot(daf)
}
