% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_genes.R
\name{search_genes}
\alias{search_genes}
\title{This function searches the  DoD_Baseline and T6_Baseline pairwise comparison results for all the search terms supplied
in the character vector; the objects DoD_Baseline_sig AND T6_Baseline_sig need to be defined in the environment for this
to work}
\usage{
search_genes(x)
}
\arguments{
\item{search_terms}{character string; one or more search terms}
}
\value{
a named list of lists containing the duplicated-cleaned gene symbols that were found using the search term
}
\description{
This function searches the  DoD_Baseline and T6_Baseline pairwise comparison results for all the search terms supplied
in the character vector; the objects DoD_Baseline_sig AND T6_Baseline_sig need to be defined in the environment for this
to work
}
