% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac63_diffcyt_boxplot.R
\name{vac63_diffcyt_boxplot}
\alias{vac63_diffcyt_boxplot}
\title{This function makes boxplots of the frequency or cell count of each cluster that's been determined to be differentially abundant using diffcyt.}
\usage{
vac63_diffcyt_boxplot(diffcyt_result, sce, counts = NULL, FDR, logFC = NULL)
}
\arguments{
\item{diffcyt_result;}{result of \code{diffcyt()} in differential abundance mode}

\item{sce;}{what SingleCellExperiment was used for diffcyt?}

\item{counts;}{logical; if \code{TRUE}, count data is displayed, if \code{FALSE} or \code{NULL} cluster frequencies are shown}

\item{FDR;}{numeric; false discovery rate used to correct for multiple testing}

\item{logFC;}{numeric; symmetrical cutoff for change to be considered relevant both up and down}

\item{shape;}{string; argument passed to aes in geom_point(). Has to be an exact match}
}
\value{
a ggplot2 object
}
\description{
This function makes boxplots of the frequency or cell count of each cluster that's been determined to be differentially abundant using diffcyt.
}
