% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_gene_heatmaps.R
\name{quick_gene_heatmaps}
\alias{quick_gene_heatmaps}
\title{This function searches the  DoD_Baseline and T6_Baseline pairwise comparison results for all the search terms supplied
in the character vector; the objects DoD_Baseline_sig AND T6_Baseline_sig need to be defined in the environment for this
to work}
\usage{
quick_gene_heatmaps(search_results)
}
\arguments{
\item{search_results}{character string; a list of lists containign gene names for making heatmaps}
}
\value{
a named list of lists containing the data frames used for plotting. The heatmaps will be save in ./figures relative to the
working directory
}
\description{
This function searches the  DoD_Baseline and T6_Baseline pairwise comparison results for all the search terms supplied
in the character vector; the objects DoD_Baseline_sig AND T6_Baseline_sig need to be defined in the environment for this
to work
}
